+++ src/main/java/de/cismet/belisEE/entity/Leuchte.java
@@ -118,14 +118,14 @@ public interface Leuchte extends Serializable, PropertyChangeListener, DocumentC
      *
      * @return  DOCUMENT ME!
      */
-    Short getLaufendeNummer();
+    Integer getLaufendeNummer();
 
     /**
      * DOCUMENT ME!
      *
      * @param  laufendeNummer  DOCUMENT ME!
      */
-    void setLaufendeNummer(final Short laufendeNummer);
+    void setLaufendeNummer(final Integer laufendeNummer);
 
     /**
      * DOCUMENT ME!
