+++ src/main/java/de/cismet/cids/custom/beans/belis/QuerschnittCustomBean.java
@@ -39,7 +39,7 @@ public class QuerschnittCustomBean extends BaseEntity implements Querschnitt {
 
     //~ Instance fields --------------------------------------------------------
 
-    private Long id;
+    private Integer id;
     private Double groesse;
 
     //~ Constructors -----------------------------------------------------------
@@ -72,13 +72,13 @@ public class QuerschnittCustomBean extends BaseEntity implements Querschnitt {
     }
 
     @Override
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
     @Override
-    public void setId(final Long id) {
-        final Long old = this.id;
+    public void setId(final Integer id) {
+        final Integer old = this.id;
         this.id = id;
         this.propertyChangeSupport.firePropertyChange(PROP__ID, old, this.id);
     }
